move api into com.github.martingaston package
