+++ api/src/main/java/com/github/martingaston/API.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import java.io.IOException;
 
 public class API {
