+++ api/src/main/java/com/github/martingaston/Encode.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
