+++ api/src/main/java/com/github/martingaston/Json.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import com.github.martingaston.tictactoe.board.Symbol;
 
 import java.util.List;
