+++ api/src/main/java/com/github/martingaston/JsonIncoming.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.martingaston.tictactoe.board.Symbol;
@@ -7,7 +9,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-class JsonOutgoing implements Json {
+class JsonIncoming implements Json {
+    private final int position;
     private final String mode;
     private final Symbol currentPlayer;
     private final boolean isActive;
@@ -15,17 +18,19 @@ class JsonOutgoing implements Json {
     private final Map<String, String> messages;
 
     @JsonCreator
-    private static JsonOutgoing from(
+    private static JsonIncoming from(
+            @JsonProperty("position") int position,
             @JsonProperty("mode") String mode,
             @JsonProperty("currentPlayer") String currentPlayer,
             @JsonProperty("isActive") boolean isActive,
             @JsonProperty("board") ArrayList<String> board,
             @JsonProperty("messages") HashMap<String, String> messages
     ) {
-        return new JsonOutgoing(mode, new Symbol(currentPlayer), isActive, board, messages);
+        return new JsonIncoming(position, mode, new Symbol(currentPlayer), isActive, board, messages);
     }
 
-    private JsonOutgoing(String mode, Symbol currentPlayer, boolean isActive, List<String> board, Map<String, String> messages) {
+    private JsonIncoming(int position, String mode, Symbol currentPlayer, boolean isActive, List<String> board, Map<String, String> messages) {
+        this.position = position;
         this.mode = mode;
         this.currentPlayer = currentPlayer;
         this.isActive = isActive;
@@ -34,11 +39,18 @@ class JsonOutgoing implements Json {
     }
 
     public static class Builder {
+        private int position;
         private String mode;
-        private Symbol currentPlayer = new Symbol("");
+        private Symbol currentPlayer;
         private boolean isActive;
         private List<String> board;
         private Map<String, String> messages;
+
+        public Builder position(int position) {
+            this.position = position;
+
+            return this;
+        }
 
         public Builder mode(String mode) {
             this.mode = mode;
@@ -70,9 +82,14 @@ class JsonOutgoing implements Json {
             return this;
         }
 
-        public JsonOutgoing build() {
-            return new JsonOutgoing(this.mode, this.currentPlayer, this.isActive, this.board, this.messages);
+        public JsonIncoming build() {
+            return new JsonIncoming(this.position, this.mode, this.currentPlayer, this.isActive, this.board, this.messages);
         }
+    }
+
+    @JsonProperty("position")
+    public int position() {
+        return this.position;
     }
 
     @JsonProperty("isActive")
