+++ api/src/main/java/com/github/martingaston/JsonOutgoing.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.martingaston.tictactoe.board.Symbol;
@@ -7,8 +9,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-class JsonIncoming implements Json {
-    private final int position;
+class JsonOutgoing implements Json {
     private final String mode;
     private final Symbol currentPlayer;
     private final boolean isActive;
@@ -16,19 +17,17 @@ class JsonIncoming implements Json {
     private final Map<String, String> messages;
 
     @JsonCreator
-    private static JsonIncoming from(
-            @JsonProperty("position") int position,
+    private static JsonOutgoing from(
             @JsonProperty("mode") String mode,
             @JsonProperty("currentPlayer") String currentPlayer,
             @JsonProperty("isActive") boolean isActive,
             @JsonProperty("board") ArrayList<String> board,
             @JsonProperty("messages") HashMap<String, String> messages
     ) {
-        return new JsonIncoming(position, mode, new Symbol(currentPlayer), isActive, board, messages);
+        return new JsonOutgoing(mode, new Symbol(currentPlayer), isActive, board, messages);
     }
 
-    private JsonIncoming(int position, String mode, Symbol currentPlayer, boolean isActive, List<String> board, Map<String, String> messages) {
-        this.position = position;
+    private JsonOutgoing(String mode, Symbol currentPlayer, boolean isActive, List<String> board, Map<String, String> messages) {
         this.mode = mode;
         this.currentPlayer = currentPlayer;
         this.isActive = isActive;
@@ -37,18 +36,11 @@ class JsonIncoming implements Json {
     }
 
     public static class Builder {
-        private int position;
         private String mode;
-        private Symbol currentPlayer;
+        private Symbol currentPlayer = new Symbol("");
         private boolean isActive;
         private List<String> board;
         private Map<String, String> messages;
-
-        public Builder position(int position) {
-            this.position = position;
-
-            return this;
-        }
 
         public Builder mode(String mode) {
             this.mode = mode;
@@ -80,14 +72,9 @@ class JsonIncoming implements Json {
             return this;
         }
 
-        public JsonIncoming build() {
-            return new JsonIncoming(this.position, this.mode, this.currentPlayer, this.isActive, this.board, this.messages);
+        public JsonOutgoing build() {
+            return new JsonOutgoing(this.mode, this.currentPlayer, this.isActive, this.board, this.messages);
         }
-    }
-
-    @JsonProperty("position")
-    public int position() {
-        return this.position;
     }
 
     @JsonProperty("isActive")
