+++ api/src/test/java/com/github/martingaston/APITest.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import org.junit.Test;
 
 import java.io.IOException;
