+++ api/src/test/java/com/github/martingaston/UpdateTest.java
@@ -1,3 +1,5 @@
+package com.github.martingaston;
+
 import com.github.martingaston.tictactoe.board.Symbol;
 import org.junit.Test;
